Выносит адаптер в utils
