+++ project/src/utils/adapter.ts
@@ -1,0 +1,25 @@
+import { Offer } from '../types/offer';
+import { BackendOffer } from '../types/backend-offer';
+
+export const adaptOfferToClient = (offer: BackendOffer): Offer => {
+  const adaptedOffer = {
+    ...offer,
+    host: {
+      avatarUrl: offer.host.avatar_url,
+      id: offer.host.id,
+      isPro: offer.host.is_pro,
+      name: offer.host.name,
+    },
+    isFavorite: offer.is_favorite,
+    isPremium: offer.is_premium,
+    maxAdults: offer.max_adults,
+    previewImage: offer.preview_image,
+  };
+
+  delete adaptedOffer['is_favorite'];
+  delete adaptedOffer['is_premium'];
+  delete adaptedOffer['max_adults'];
+  delete adaptedOffer['preview_image'];
+
+  return adaptedOffer as Offer;
+};
